Reorganize tests tree
This PR reorganizes the test tree to prepare for future tests.

The organization is as follows:
- tests/           # all test code
  - garage/        # unit tests for modules. mirrors package tree.
  - fixtures/      # test fixtures -- code used to build good tests
  - benchmarks/    # benchmark tests
