+++ setup.cfg
@@ -68,7 +68,7 @@ select = E101,W191,E201,E202,E241,E211,E221,W291,E702,E714,E125,E128,E501,
 # style, this rule is ignored.
 ignore = W503
 import-order-style = google
-application-import-names = sandbox,garage,examples,contrib
+application-import-names = sandbox,garage,examples,contrib,tests
 
 [pylint]
 #[MESSAGES CONTROL]
