--- tests/envs/dummy/__init__.py
@@ -1,3 +1,0 @@
-from tests.envs.dummy.dummy_box_env import DummyBoxEnv
-from tests.envs.dummy.dummy_dict_env import DummyDictEnv
-from tests.envs.dummy.dummy_discrete_env import DummyDiscreteEnv
