+++ tests/fixtures/envs/dummy/__init__.py
@@ -1,0 +1,3 @@
+from tests.fixtures.envs.dummy.dummy_box_env import DummyBoxEnv
+from tests.fixtures.envs.dummy.dummy_dict_env import DummyDictEnv
+from tests.fixtures.envs.dummy.dummy_discrete_env import DummyDiscreteEnv
