+++ tests/garage/tf/misc/test_tensor_utils.py
