+++ tests/garage/tf/policies/test_categorical_policies.py
