+++ tests/garage/theano/algos/test_algos.py
