+++ tests/garage/theano/algos/test_ddpg.py
