+++ tests/garage/theano/algos/test_trpo.py
