+++ tests/garage/theano/core/test_instrument.py
