+++ tests/garage/theano/core/test_networks.py
