+++ tests/garage/theano/policies/test_adaptive_std.py
