+++ tests/garage/theano/spaces/test_spaces.py
